# packages/openapi/skill-registry.openapi.yaml
openapi: 3.1.0
info:
  title: ManagedSkillHub API
  description: Self-hosted skill registry for AI agents.
  version: 0.1.0
servers:
- url: http://localhost:3040
  description: Local API server
...
paths:
...
  /proposals/check-duplicate:
    post:
      operationId: checkDuplicateProposal
      tags:
      - Proposals
      summary: Pre-submission duplicate and similarity check
      description: 'May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or agentOidc
        when PROPOSAL_AUTH_MODE=oidc.

        Checks whether an identical or similar proposal/skill already exists before
        creating a proposal.

        Agents should call this endpoint before submitting. Submission is not blocked
        by matches;

        the response only provides transparency about existing content and skill id
        collisions.

        '
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckDuplicateProposalRequest'
      responses:
        '200':
          description: Duplicate check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckDuplicateProposalResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /proposals:
    post:
      operationId: submitProposal
      tags:
      - Proposals
      summary: Submit a new proposal
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or agentOidc
        when PROPOSAL_AUTH_MODE=oidc.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitProposalRequest'
      responses:
        '201':
          description: Proposal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalSubmissionResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /proposals/notice:
    get:
      operationId: getProposalNotice
      tags:
      - Proposals
      summary: Show whether new proposals are waiting for admin review
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or agentOidc
        when PROPOSAL_AUTH_MODE=oidc.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      responses:
        '200':
          description: Aggregated proposal notice
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalNoticeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
  /proposals/{proposalId}/status:
    get:
      operationId: getProposalStatus
      tags:
      - Proposals
      summary: Public proposal status lookup
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or any
        allowed agentOidc proposal principal when PROPOSAL_AUTH_MODE=oidc. Knowing
        the UUID is insufficient without the configured authentication.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Public proposal status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalPublicStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /proposals/{proposalId}:
    patch:
      operationId: updateOpenProposalMetadata
      tags:
      - Proposals
      summary: Update metadata for an open proposal upload
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or the
        owning agentOidc principal when oidc. Only proposals still in `in_upload`
        can be updated through this public submitter path. Agents should use this
        after submitter-side post-checks when title, description, category, tags,
        capabilities, or entrypoint need correction before finalization.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalMetadataUpdateRequest'
      responses:
        '200':
          description: Proposal metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalMetadataUpdateResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: deleteOpenProposalUpload
      tags:
      - Proposals
      summary: Abort and delete an open proposal upload
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or the
        owning agentOidc principal when oidc. Only proposals still in `in_upload`
        can be deleted through this submitter path. After finalize-upload, public
        proposal deletion is blocked.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Open proposal upload deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /proposals/{proposalId}/files:
    post:
      operationId: attachProposalFile
      tags:
      - Proposals
      summary: Attach a file to a proposal
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or the
        owning agentOidc principal when oidc. Attach one proposal artifact. When the
        file belongs in a subfolder, send multipart field `path` with the relative
        in-package path such as `scripts/build.py`. While the proposal is still `in_upload`,
        sending the same relative path again replaces that file so agents can apply
        post-check corrections without creating another proposal.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                path:
                  type: string
                  description: Optional relative in-package path. Use this to preserve
                    subfolders like scripts/, templates/, docs/, or assets/ during
                    upload.
      responses:
        '200':
          description: File attached
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalAttachmentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /proposals/{proposalId}/validate-upload:
    post:
      operationId: validateProposalUpload
      tags:
      - Proposals
      summary: Validate an open proposal upload without finalizing
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or the
        owning agentOidc principal when oidc. Checks the currently uploaded package
        while the proposal is still `in_upload`, returns all package-reference findings,
        and does not finalize the upload, run extraction, run judgement, or change
        proposal status. Agents should call this before finalize-upload and apply
        corrections through PATCH /proposals/{proposalId} or POST /proposals/{proposalId}/files
        upserts.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Upload validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalUploadValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /proposals/{proposalId}/finalize-upload:
    post:
      operationId: finalizeProposalUpload
      tags:
      - Proposals
      summary: Finalize an open proposal upload and start judgement
      description: May require proposalBearer when PROPOSAL_AUTH_MODE=bearer or the
        owning agentOidc principal when oidc.
      security:
      - proposalBearer: []
      - agentSession: []
      - agentOidc:
        - managedskillhub:proposals
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Upload finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalFinalizeUploadResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /admin/proposals:
    get:
      operationId: adminListProposals
      tags:
      - Admin Proposals
      summary: List proposals for review
      security:
      - adminSession: []
      parameters:
      - name: skillId
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: string
          description: Optional comma-separated list of proposal states.
      responses:
        '200':
          description: Proposal list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /admin/proposals/{proposalId}:
    get:
      operationId: adminGetProposal
      tags:
      - Admin Proposals
      summary: Get proposal details
      security:
      - adminSession: []
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Proposal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      operationId: adminUpdateProposal
      tags:
      - Admin Proposals
      summary: Update proposal metadata before conversion
      security:
      - adminSession: []
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalAdminUpdateRequest'
      responses:
        '200':
          description: Proposal details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
    delete:
      operationId: adminDeleteProposal
      tags:
      - Admin Proposals
      summary: Delete an abandoned open proposal upload
      security:
      - adminSession: []
      parameters:
      - name: proposalId
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: skill_hub_session
    publicReadBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for published skill consumption
        when PUBLIC_READ_AUTH_MODE=bearer.
    proposalBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for proposal duplicate check,
        submit, upload, finalization, notices, and status when PROPOSAL_AUTH_MODE=bearer.
    discoveryBearer:
      type: http
      scheme: bearer
      description: Runtime-configurable bearer token for discovery/contract endpoints
        when DISCOVERY_AUTH_MODE=bearer.
    agentOidc:
      type: openIdConnect
      openIdConnectUrl: https://auth.example.invalid/application/o/managedskillhub-agent-device/.well-known/openid-configuration
      description: Runtime Authentik access token. The concrete issuer, client ID,
        scopes, Device Authorization endpoint, token endpoint, and OpenID configuration
        URL must be read from /discover; this static URL is intentionally only a contract
        placeholder.
      x-device-authorization:
        flow: device_code
        runtimeMetadataPath: /discover
        verificationLinkField: verification_uri_complete
        deviceCodeIsSecret: true
    agentSession:
      type: apiKey
      in: header
      name: Authorization
      description: 'Short-lived agent session code sent as "Authorization: AgentSession
        <code>". Created through POST /agent-sessions after proving possession of
        the relevant area bearer tokens.'
  responses:
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Conflict error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Authenticated actor is not allowed to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimitError:
      description: Rate limit exceeded
      headers:
        Retry-After:
          description: Seconds until the current rate-limit window resets.
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    SubmitProposalRequest:
      type: object
      required:
      - title
      - description
      - category
      properties:
        skillId:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        entrypoint:
          type: string
...
    CheckDuplicateProposalRequest:
      type: object
      required:
      - title
      - description
      - category
      properties:
        skillId:
          type: string
          description: Optional target skill id. If it already exists, conversion
            would create a new draft version.
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        entrypoint:
          type: string
        files:
          type: array
          description: Optional file fingerprints. If sha256 values are provided,
            an exact content digest can be computed and checked against existing proposals/skills
            without uploading the full content.
          items:
            type: object
            required:
            - path
            properties:
              path:
                type: string
              sha256:
                type: string
                nullable: true
...
    CheckDuplicateProposalResponse:
      type: object
      required:
      - submittedContentDigest
      - exactDuplicateProposalId
      - exactDuplicateSkillId
      - similarMatches
      - skillIdCollision
      - resolutionOptions
      - note
      properties:
        submittedContentDigest:
          type: string
          nullable: true
          description: Computed content digest if file fingerprints were provided,
            otherwise null.
        exactDuplicateProposalId:
          type: string
          nullable: true
        exactDuplicateSkillId:
          type: string
          nullable: true
        similarMatches:
          type: array
          items:
            $ref: '#/components/schemas/CheckDuplicateProposalMatch'
        skillIdCollision:
          type: object
          required:
          - exists
          - existingSkillId
          - note
          properties:
            exists:
              type: boolean
            existingSkillId:
              type: string
              nullable: true
            note:
              type: string
        resolutionOptions:
          type: array
          items:
            $ref: '#/components/schemas/CheckDuplicateProposalResolutionOption'
          description: Actions the submitter can choose from based on duplicate/skill-id
            collision findings.
        note:
          type: string
    ProposalSubmissionResponse:
      type: object
      required:
      - id
      - message
      - statusUrl
      - checkUrl
      - finalizeUploadUrl
      properties:
        id:
          type: string
        message:
          type: string
          description: Human-readable confirmation that the upload was opened and
            must still be finalized by the submitter.
        statusUrl:
          type: string
          description: Absolute or prefix-aware URL to poll the public proposal status.
        checkUrl:
          type: string
          description: Alias for statusUrl, provided for convenience.
        finalizeUploadUrl:
          type: string
          description: Absolute or prefix-aware URL to explicitly finalize the upload
            after all files were attached.
    ProposalMetadataUpdateRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        entrypoint:
          type: string
          nullable: true
    ProposalMetadataUpdateResponse:
      type: object
      required:
      - id
      - status
      - title
      - description
      - category
      - tags
      - capabilities
      - entrypoint
      properties:
        id:
          type: string
        status:
          type: string
          enum:
          - in_upload
        title:
          type: string
        description:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        capabilities:
          type: array
          items:
            type: string
        entrypoint:
          type: string
          nullable: true
...
    ProposalNoticeResponse:
      type: object
      required:
      - hasNewProposals
      - totalPending
      properties:
        hasNewProposals:
          type: boolean
        totalPending:
          type: integer
...